requirements.txt
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
fastapi==0.101.1
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
python-logstash==0.4.8
PyYAML==6.0.1
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
<<<<<<< HEAD
=======

>>>>>>> development
